Update Character Creation
Changed Weight Storage Data Types.
Added Weight Help Test.

+++ DnD35EDMTools/Helpers/WeightHelper.cs
@@ -21,8 +21,8 @@ public static class WeightHelper
         }
     }
     
-    public static int GetWeight(int weight, string size)
+    public static double GetWeight(double weight, string size)
     {
-        return (int)(weight * GetMultiplier(size));
+        return (double)(weight * GetMultiplier(size));
     }
 }
